Refactor methods for search books
