+++ src/main/java/com/epam/tishkin/models/Author.java
@@ -12,13 +12,9 @@ import java.util.List;
 @XmlRootElement
 public class Author implements Serializable {
     private static final long serialVersionUID = 785478548L;
-
     @Id
     @XmlElement
     private String name;
-    @OneToMany(mappedBy = "author", orphanRemoval = true, fetch = FetchType.EAGER)
-    @XmlElement
-    private List<Book> book;
 
     public Author() {
     }
@@ -33,14 +29,6 @@ public class Author implements Serializable {
 
     public void setName(String name) {
         this.name = name;
-    }
-
-    public List<Book> getBook() {
-        return book;
-    }
-
-    public void setBook(List<Book> book) {
-        this.book = book;
     }
 
     @Override
